Reorganized EVT serializable code.
